+++ src/components/Chip.jsx
@@ -1,18 +1,25 @@
-import { IconTrash } from "@tabler/icons-react";
+/* eslint-disable react/prop-types */
+import { IconTrash } from "@tabler/icons-react"
+import "./Chip.css"
+import { useGiftsStore } from "../store/useGiftsStore"
+
 
 export function Chip({el}){
+    const deleteGift=useGiftsStore(state=>state.deleteGift);
     return (
-            <li key={el.name}>
-                <div className="leftSideChip">
-                    <img src={el.link} alt={el.name} />
-                    <div className="chipInfo">
-                        <p>{el.name}</p>
-                        <p>Cantidad: {el.qty}</p>
-                    </div>
+        <li>
+            <div className="leftSideChip">
+                <img src={el.link} alt={el.name} />
+                <div className="chipInfo">
+                    <p>{el.name}</p>
+                    <p>Cantidad: {el.qty}</p>
                 </div>
-                <>
-                    <button className="deleteButton"><IconTrash /></button>
-                </>
-            </li>
+            </div>
+            <>
+                <button className="deleteButton" onClick={()=>deleteGift(el)}>
+                    <IconTrash />
+                </button>
+            </>
+        </li>
     )
 }
