--- src/components/Modal.jsx
@@ -1,11 +1,0 @@
-import "./Modal.css";
-
-export function Modal({children}){
-    return (
-        <div className="modalOverlay">
-            <section className="modalContainer">
-                {children}
-            </section>
-        </div>
-    )
-}
