--- src/hooks/useModal.js
@@ -1,7 +1,0 @@
-import { useState } from "react"
-
-export const useModal=()=>{
-    const [modals, setModals]=useState({type:null})
-
-    return {modals, setModals}
-}
